rename db.js for dbMain.js feat: freeing port before starting dev
