+++ models/dbCompatibility.js
@@ -1,0 +1,15 @@
+import mysql from 'mysql2';
+
+const pool = mysql.createPool({
+    host: process.env.MYSQL_HOST,
+    user: process.env.MYSQL_USER,
+    password: process.env.MYSQL_PASSWORD,
+    database: process.env.MYSQL_DATABASE,
+}).promise();
+
+
+async function checkCpuAndMoboSocketComp(cpuId, moboId) {
+    const [cpuSocket] = await pool.query("select socket from cpu where id = ?", [cpuId]);
+    const [moboSocket] = await pool.query("select socket from mobo where id = ?", [moboId]);
+    return cpuSocket[0].socket === moboSocket[0].socket;
+}
