+++ views/products/motherboard.ejs
@@ -48,8 +48,8 @@
                     </td>
                     <td><%= mobo.model %></td>
                     <td><%= mobo.socket %></td>
-                    <td><%= mobo.chipset %> GHz</td>
-                    <td><%= mobo.form_factor %> GHz</td>
+                    <td><%= mobo.chipset %></td>
+                    <td><%= mobo.form_factor %></td>
                     <td><%= mobo.max_memory %></td>
                     <td><%= mobo.memory_slots %></td>
                     <td>
